+++ src/main/java/park/bumsiku/controller/PublicController.java
@@ -6,9 +6,9 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.view.RedirectView;
 import park.bumsiku.domain.dto.request.CommentRequest;
 import park.bumsiku.domain.dto.response.*;
-import park.bumsiku.log.aop.LogExecutionTime;
+import park.bumsiku.utils.monitoring.LogExecutionTime;
 import park.bumsiku.service.PublicService;
-import park.bumsiku.utils.ArgumentValidator;
+import park.bumsiku.utils.integration.ArgumentValidator;
 
 import java.util.List;
