+++ src/main/java/park/bumsiku/utils/GlobalExceptionHandler.java
@@ -13,6 +13,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 import org.springframework.web.servlet.resource.NoResourceFoundException;
 import park.bumsiku.domain.dto.response.Response;
+import park.bumsiku.utils.validation.DiscordWebhookCreator;
 
 import java.util.NoSuchElementException;
