+++ src/main/java/park/bumsiku/utils/monitoring/LoggingAspect.java
@@ -1,4 +1,4 @@
-package park.bumsiku.log.aop;
+package park.bumsiku.utils.monitoring;
 
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
@@ -11,7 +11,7 @@ public class LoggingAspect extends AbstractLoggingAspect {
 
     private static final long SLOW_METHOD_THRESHOLD_MS = 500;
 
-    @Around("@annotation(park.bumsiku.log.aop.LogExecutionTime)")
+    @Around("@annotation(park.bumsiku.utils.monitoring.LogExecutionTime)")
     public Object logExecutionTime(ProceedingJoinPoint joinPoint) throws Throwable {
         return super.logExecutionTime(joinPoint, "method execution", SLOW_METHOD_THRESHOLD_MS);
     }
