+++ src/test/java/park/bumsiku/controller/PublicControllerTest.java
@@ -14,8 +14,8 @@ import park.bumsiku.config.SecurityConfig;
 import park.bumsiku.domain.dto.request.CommentRequest;
 import park.bumsiku.domain.dto.response.*;
 import park.bumsiku.service.PublicService;
-import park.bumsiku.utils.ArgumentValidator;
-import park.bumsiku.utils.DiscordWebhookCreator;
+import park.bumsiku.utils.integration.ArgumentValidator;
+import park.bumsiku.utils.validation.DiscordWebhookCreator;
 
 import java.time.LocalDateTime;
 import java.util.Arrays;
