+++ src/test/java/park/bumsiku/service/PublicServiceTest.java
@@ -16,7 +16,7 @@ import park.bumsiku.domain.entity.Post;
 import park.bumsiku.repository.CategoryRepository;
 import park.bumsiku.repository.CommentRepository;
 import park.bumsiku.repository.PostRepository;
-import park.bumsiku.utils.DiscordWebhookCreator;
+import park.bumsiku.utils.validation.DiscordWebhookCreator;
 
 import java.time.LocalDateTime;
 import java.util.List;
