+++ src/test/resources/logback-test.xml
@@ -12,7 +12,7 @@
         <appender-ref ref="CONSOLE"/>
     </logger>
 
-    <logger name="park.bumsiku.log.LoggingFilter" level="INFO" additivity="false">
+    <logger name="park.bumsiku.utils.monitoring.LoggingFilter" level="INFO" additivity="false">
         <appender-ref ref="CONSOLE"/>
     </logger>
